educational-platform-backend/chatbot-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-core
langchain-openai
langchain-ollama
pydantic